Start consolidating TypeScript and Flow into a common parser
There were some completely copied functions between the two, and probably other
code can be combined later. It may also be reasonable to have a simpler parser
that parses a superset of both, with only a few special cases, but we'll see.
